Converting dynamoDB to typescript
